Feat: Bar 컴포넌트 내부 요소 하위 컴포넌트로 분리 및 로직 이동
