--- src/pages/PostPage/UserPost/Bar.module.scss
@@ -1,126 +1,0 @@
-@import '../../../styles/index';
-
-.barArea {
-  display: flex;
-  justify-content: space-between;
-  width: 62.8%;
-}
-
-.name {
-  text-align: left;
-  letter-spacing: -0.01em;
-
-  @include text-3xl;
-}
-
-.barContents {
-  position: relative;
-  display: flex;
-  gap: 1.8rem;
-  align-items: center;
-}
-
-.writeArea {
-  display: flex;
-  gap: 0.6rem;
-  align-items: center;
-}
-
-.profiles {
-}
-
-.writingCount {
-  display: flex;
-  font-weight: 400;
-  text-align: left;
-
-  @include text-lg;
-}
-
-.count {
-  @include font-bold;
-}
-
-.emojiArea {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-  border-right: 1px;
-  border-left: 1px;
-}
-
-.reactionArea {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.reaction {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.reactions {
-  position: absolute;
-  top: 3.1rem;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.6rem;
-  align-content: center;
-  width: 16.8rem;
-  padding: 1.5rem;
-  background-color: $white;
-  border: 1px solid $gray-200;
-
-  @include rounded-lg;
-}
-
-.arrowDown {
-  width: 0.8rem;
-}
-
-.share {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 0.4rem;
-  width: 3.5rem;
-  height: 2.3rem;
-  border: 1px solid $gray-200;
-
-  @include rounded-sm;
-}
-
-.shareImage {
-  display: flex;
-  flex-flow: row wrap;
-  align-items: center;
-  justify-content: center;
-  width: 3.5rem;
-  height: 2.3rem;
-  border: 1px;
-
-  @include rounded-sm;
-}
-
-.shareArea {
-  background-color: $white;
-  border: 1px solid $gray-200;
-
-  @include rounded-lg;
-}
-
-.shareSns {
-  position: relative;
-  display: flex;
-  align-items: center;
-  width: 8.6rem;
-  height: 3.1rem;
-}
-
-.shareSns:hover {
-  background-color: $gray-100;
-}
-
-.shareSnsText {
-  margin-left: 0.1rem;
-}
