+++ src/pages/PostPage/UserPost/Bar/Bar.module.scss
@@ -1,0 +1,21 @@
+@import '../../../../styles/index';
+
+.barArea {
+  display: flex;
+  justify-content: space-between;
+  width: 62.8%;
+}
+
+.name {
+  text-align: left;
+  letter-spacing: -0.01em;
+
+  @include text-3xl;
+}
+
+.barContents {
+  position: relative;
+  display: flex;
+  gap: 1.8rem;
+  align-items: center;
+}
