+++ src/pages/PostPage/UserPost/Bar/Share.module.scss
@@ -1,0 +1,47 @@
+@import '../../../../styles/index';
+
+.share {
+  display: flex;
+  flex-flow: row wrap;
+  gap: 0.4rem;
+  width: 3.5rem;
+  height: 2.3rem;
+  border: 1px solid $gray-200;
+
+  @include rounded-sm;
+}
+
+.shareImage {
+  display: flex;
+  flex-flow: row wrap;
+  align-items: center;
+  justify-content: center;
+  width: 3.5rem;
+  height: 2.3rem;
+  border: 1px;
+
+  @include rounded-sm;
+}
+
+.shareArea {
+  background-color: $white;
+  border: 1px solid $gray-200;
+
+  @include rounded-lg;
+}
+
+.shareSns {
+  position: relative;
+  display: flex;
+  align-items: center;
+  width: 8.6rem;
+  height: 3.1rem;
+}
+
+.shareSns:hover {
+  background-color: $gray-100;
+}
+
+.shareSnsText {
+  margin-left: 0.1rem;
+}
