+++ src/pages/PostPage/UserPost/Bar/WritingCount.module.scss
@@ -1,0 +1,19 @@
+@import '../../../../styles/index';
+.writeArea {
+  display: flex;
+  gap: 0.6rem;
+  align-items: center;
+  border: 1px solid black;
+}
+
+.writingCount {
+  display: flex;
+  font-weight: 400;
+  text-align: left;
+
+  @include text-lg;
+}
+
+.count {
+  @include font-bold;
+}
