+++ src/pages/PostPage/UserPost/Post.jsx
@@ -1,6 +1,4 @@
-import Bar from './Bar';
-// import Badge from '../../components/Badge/Badge';
-// import BadgeEmoji from '../../components/Badge/BadgeEmoji';
+import Bar from './Bar/Bar';
 import css from './Post.module.scss';
 const Post = () => {
   return (
